backend/requirements.txt
flask
Flask-SQLAlchemy
flask-login
werkzeug
python-dotenv
sqlalchemy